Remove Parcel.equals() and make it interal to PureParcel
