+++ packages/parcels-react/src/__test__/PureParcel-test.js
@@ -8,6 +8,7 @@ configure({adapter: new Adapter()});
 
 import {shallow} from 'enzyme';
 import PureParcel from '../PureParcel';
+import PureParcelEquals from '../PureParcelEquals';
 import Parcel from 'parcels';
 
 test('PureParcel should pass a *value equivalent* parcel to children', tt => {
@@ -16,7 +17,7 @@ test('PureParcel should pass a *value equivalent* parcel to children', tt => {
 
     let wrapper = shallow(<PureParcel parcel={parcel}>
         {(pp) => {
-            tt.true(pp.equals(parcel));
+            tt.true(PureParcelEquals(pp, parcel));
         }}
     </PureParcel>);
 });
@@ -49,9 +50,9 @@ test('PureParcel should pass a NEW *value equivalent* parcel to children when pr
     let wrapper = shallow(<PureParcel parcel={parcel}>
         {(pp) => {
             if(renders === 0) {
-                tt.true(pp.equals(parcel));
+                tt.true(PureParcelEquals(pp, parcel));
             } else if(renders === 1) {
-                tt.true(pp.equals(parcel2));
+                tt.true(PureParcelEquals(pp, parcel2));
             }
             renders++;
         }}
