+++ packages/parcels/src/parcel/__test__/ValueParcelMethods-test.js
@@ -404,30 +404,6 @@ test('Parcel.setChangeRequestMeta() should set change request meta', (tt: Object
     });
 });
 
-test('Parcel.equals() should compare two parcels data', (tt: Object) => {
-    var child = {};
-    var parcelCreator = (merge = {}) => {
-        let p = new Parcel();
-        p._parcelData = {
-            value: 123,
-            meta: {
-                abc: 123,
-                def: 456
-            },
-            key: "a",
-            child,
-            ...merge
-        };
-        return p;
-    };
-
-    tt.true(parcelCreator().equals(parcelCreator()), 'parcel equals self');
-    tt.false(parcelCreator().equals(parcelCreator({value: 456})), 'parcel doesnt equals different value');
-    tt.false(parcelCreator().equals(parcelCreator({meta: {abc: 123}})), 'parcel doesnt equals different meta contents');
-    tt.false(parcelCreator().equals(parcelCreator({child: {}})), 'parcel doesnt equals different child');
-    tt.false(parcelCreator().equals(parcelCreator({key: "b"})), 'parcel doesnt equals different key');
-});
-
 test('Parcel.hasDispatched() should say if a parcel has dispatched from the current parcels path location', (tt: Object) => {
     tt.plan(6);
